requirements.txt
crewai
crewai_tools
langchain-core
langchain
langchain-community
python_dotenv
groq
langchain-groq
asyncio
matplotlib
torch

# requirements file updated to version with more state data

#Python 3.11.1
# sudo apt-get install python3-dev graphviz libgraphviz-dev pkg-config

#Models Used:
#OpenAI:
# - gpt-3.5, gpt-4o : Standard Key
# - tavily search tool : Joe has the key

python-dotenv~=1.0.1
tavily-python==0.3.3
duckduckgo_search==5.3.1b1
beautifulsoup4==4.12.3
openai==1.30.1
langgraph==0.0.53
langchain_core==0.2.0
langchain_openai==0.1.7
langchain-community==0.2.0
langchain==0.2.0
aiosqlite==0.20.0
pygments==2.17.2
#pygraphviz
gradio==4.31.3
transformers==4.41.2
accelerate==0.31.0
pyautogen==0.2.25
chess==1.10.0
matplotlib
numpy
pandas
yfinance
autogen-agentchat
pyautogen[retrievechat]
markdownify
sentence_transformers
llama-index-vector-stores-chroma
